backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Cache and Sessions
redis==5.0.1

# WebSocket
websockets==12.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Machine Learning
lightgbm==4.1.0
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0

# Network Monitoring
scapy==2.5.0

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
